modified:   uart_lepteto/ADC.c 	modified:   uart_lepteto/ADC.h 	new file:   uart_lepteto/Makefile 	modified:   uart_lepteto/crc.c 	modified:   uart_lepteto/crc.h 	new file:   uart_lepteto/git/MSP430_UART 	modified:   uart_lepteto/init_uart.c 	modified:   uart_lepteto/init_uart.h 	new file:   uart_lepteto/main.c 	modified:   uart_lepteto/reading.c 	modified:   uart_lepteto/reading.h 	modified:   uart_lepteto/uart_lepteto.ino 	modified:   uart_lepteto/writing.c 	modified:   uart_lepteto/writing.h 	modified:   uart_test_ok/uart_test_ok.ino

+++ uart_lepteto/writing.h
@@ -1,5 +1,6 @@
 #ifndef WRITING_H_INCLUDED
 #define WRITING_H_INCLUDED
 
-void sending(unsigned char address, unsigned char cmd,char *data, uint16_t dLen);
+void sending(unsigned char address, unsigned char cmd,char *data, int dLen);
 #endif
+
